% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_get_flux_info.R
\name{get_cycle_edge_info}
\alias{get_cycle_edge_info}
\title{get_cycle_edge_info function}
\usage{
get_cycle_edge_info(
  input_net_file,
  output_path,
  res_path,
  input_tumor_name,
  input_tumor_data,
  input_normal_data,
  prm_1 = 0.05,
  prm_2 = 1,
  prm_3 = 1,
  prm_4 = 2
)
}
\arguments{
\item{get_cycle_edge_info}{input_net_file, output_path, res_path, input_tumor_name, input_tumor_data, input_normal_data
prm_1=0.05
prm_2=1
prm_3=1
prm_4=2}
}
\description{
This function for cycle get_cycle_edge_info
}
\examples{
get_cycle_edge_info(input_net_file, output_path, res_path, input_tumor_name, input_tumor_data, input_normal_data, prm_1=0.05, prm_2=1, prm_3=1, prm_4=2)

DE_cof:
(gene pvalue < prm_1) & (abs(gene fc) > prm_2)
-> DE num ++
-> DE_cof = DE num/gene num

prm_3=1
For an edge, if any gene on the edge is up, the edge is up
prm_3=2
For an edge, if the gene with the largest mean of tumor value is up, then the edge is up

prm_4=1
For an edge, if any gene on the edge is gap, the edge is gap
prm_4=2
For an edge, if the gene with the largest mean of tumor value is gap, then the edge is gap
}
\keyword{get_cycle_edge_info}
