% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_get_graph.R
\name{get_cycle_shift}
\alias{get_cycle_shift}
\title{get_cycle_shift function}
\usage{
get_cycle_shift(
  input_net_file,
  output_path,
  res_path,
  graph_path,
  input_tumor_name,
  input_tumor_data,
  input_normal_data
)
}
\arguments{
\item{get_cycle_shift}{input_net_file, output_path, res_path, graph_path, input_tumor_name, input_tumor_data, input_normal_data}
}
\description{
This function for cycle get_cycle_shift
}
\examples{
get_cycle_shift(input_net_file, output_path, res_path, graph_path, input_tumor_name, input_tumor_data, input_normal_data)

}
\keyword{get_cycle_shift}
