% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_get_stat.R
\name{get_freq_stat}
\alias{get_freq_stat}
\title{get_freq_stat function}
\usage{
get_freq_stat(
  input_net_file,
  output_path,
  res_path,
  graph_path,
  input_tumor_name,
  input_tumor_data,
  input_normal_data
)
}
\arguments{
\item{get_freq_stat}{input_net_file, output_path, res_path, input_tumor_name, input_tumor_data, input_normal_data}
}
\description{
This function for cycle get_freq_stat
}
\examples{
get_freq_stat(input_net_file, output_path, res_path, input_tumor_name, input_tumor_data, input_normal_data)

}
\keyword{get_freq_stat}
