% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_get_graph.R
\name{get_graph_single_cycle_degree_1}
\alias{get_graph_single_cycle_degree_1}
\title{get_graph_single_cycle_degree_1 function}
\usage{
get_graph_single_cycle_degree_1(
  input_net_file,
  output_path,
  res_path,
  graph_path,
  input_tumor_name,
  input_tumor_data,
  input_normal_data
)
}
\arguments{
\item{get_graph_single_cycle_degree_1}{input_net_file, output_path, res_path, graph_path, input_tumor_name, input_tumor_data, input_normal_data}
}
\description{
This function for cycle get_graph_single_cycle_degree_1
}
\examples{
get_graph_single_cycle_degree_1(input_net_file, output_path, res_path, graph_path, input_tumor_name, input_tumor_data, input_normal_data)

}
\keyword{get_graph_single_cycle_degree_1}
