% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_get_graph.R
\name{get_graph_topology_struct}
\alias{get_graph_topology_struct}
\title{get_graph_topology_struct function}
\usage{
get_graph_topology_struct(
  input_net_file,
  output_path,
  res_path,
  graph_path,
  input_tumor_name,
  input_pathway_name,
  input_tumor_data,
  input_normal_data
)
}
\arguments{
\item{get_graph_topology_struct}{input_net_file, output_path, res_path, graph_path, input_tumor_name, input_pathway_name, input_tumor_data, input_normal_data}
}
\description{
This function for cycle get_graph_topology_struct
}
\examples{
get_graph_topology_struct(input_net_file, output_path, res_path, graph_path, input_tumor_name, input_pathway_name, input_tumor_data, input_normal_data)

}
\keyword{get_graph_topology_struct}
