% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_get_topo_info.R
\name{get_struct_sort_info}
\alias{get_struct_sort_info}
\title{get_struct_sort_info function}
\usage{
get_struct_sort_info(
  input_net_file,
  output_path,
  res_path,
  hub_prm_1 = 5,
  hub_prm_2 = 5,
  hub_prm_3 = 3,
  web_prm_1 = 0.18,
  web_prm_2 = 3,
  web_prm_3 = 7,
  web_prm_4 = 2,
  web_prm_5 = 10,
  idv_prm_1 = 0.16,
  idv_prm_2 = 1,
  idv_prm_3 = 2,
  idv_prm_4 = 20,
  idv_prm_5 = 0.56,
  idv_prm_6 = 20
)
}
\arguments{
\item{get_struct_sort_info}{input_net_file output_path res_path
hub_prm_1 = 5,
hub_prm_2 = 5,
hub_prm_3 = 3,
web_prm_1 = .18,
web_prm_2 = 3,
web_prm_3 = 7,
web_prm_4 = 2,
web_prm_5 = 10,
idv_prm_1 = .16,
idv_prm_2 = 1,
idv_prm_3 = 2,
idv_prm_4 = 20,
idv_prm_5 = .56,
idv_prm_6 = 20,}
}
\description{
This function for cycle get_struct_sort_info
}
\examples{
get_struct_sort_info()


}
\keyword{get_struct_sort_info}
