% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_get_flux_info.R
\name{get_subnet_edge_info}
\alias{get_subnet_edge_info}
\title{get_subnet_edge_info function}
\usage{
get_subnet_edge_info(
  input_net_file,
  output_path,
  res_path,
  input_tumor_name,
  input_tumor_data,
  input_normal_data,
  prm_1 = 0.05,
  prm_2 = 1,
  prm_3 = 1,
  prm_4 = 2
)
}
\arguments{
\item{get_subnet_edge_info}{input_net_file, output_path, res_path, input_tumor_name, input_tumor_data, input_normal_data}
}
\description{
This function for cycle get_subnet_edge_info
}
\examples{
get_subnet_edge_info(input_net_file, output_path, res_path, input_tumor_name, input_tumor_data, input_normal_data)

}
\keyword{get_subnet_edge_info}
