% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_get_topo_info.R
\name{get_topology_info}
\alias{get_topology_info}
\title{get_topology_info function}
\usage{
get_topology_info(input_net_file, output_path, res_path)
}
\arguments{
\item{get_topology_info}{input_net_file output_path, res_path}
}
\description{
This function for cycle get_topology_info
}
\examples{
get_topology_info(input_net_file, output_path, res_path)


}
\keyword{get_topology_info}
