% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getBasicCycle}
\alias{getBasicCycle}
\title{getBasicCycle function}
\usage{
getBasicCycle(net)
}
\description{
Build a map based on net, search for basic cycles, and calculate the expression information of these cycles
}
\examples{
getBasicCycle(net)


}
\keyword{getBasicCycle}
