% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getCycleFlux}
\alias{getCycleFlux}
\title{getCycleFlux function}
\usage{
getCycleFlux(
  input_net_file,
  input_stat_gene_file,
  input_deg_gene_file,
  prm_1 = 0.05,
  prm_2 = 1,
  prm_3 = 1
)
}
\arguments{
\item{prm_1&prm_2:}{FOR DE_cof
(gene pvalue < prm_1) & (abs(gene fc) > prm_2)
-> DE num ++
-> DE_cof = DE num/gene num}

\item{prm_3:}{IF prm_3=1
For an edge, if any gene on the edge is up, the edge is up
IF prm_3=2
For an edge, if the gene with the largest mean of tumor value is up, then the edge is up}

\item{Default:}{prm_1=0.05
prm_2=1
prm_3=1}
}
\description{
This function for cycle shift situation
}
\examples{
getCycleFlux(input_net_file, input_stat_gene_file, input_deg_gene_file, 0.05, 1, 1)

}
\keyword{getCycleFlux}
