% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_get_flux_info.R
\name{get_cycle_edge_info}
\alias{get_cycle_edge_info}
\title{get_cycle_edge_info function}
\usage{
get_cycle_edge_info(
  input_net_file,
  output_path,
  res_path,
  input_tumor_name,
  input_tumor_data,
  input_normal_data
)
}
\arguments{
\item{get_cycle_edge_info}{input_net_file, output_path, res_path, input_tumor_name, input_tumor_data, input_normal_data}
}
\description{
This function for cycle get_cycle_edge_info
}
\examples{
get_cycle_edge_info(input_net_file, output_path, res_path, input_tumor_name, input_tumor_data, input_normal_data)

}
\keyword{get_cycle_edge_info}
