% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_get_flux_info.R
\name{get_cycle_edge_info}
\alias{get_cycle_edge_info}
\title{get_cycle_edge_info function}
\usage{
get_cycle_edge_info(
  input_net_file,
  output_path,
  res_path,
  input_tumor_name,
  input_tumor_data,
  input_normal_data,
  par_up1,
  par_up2,
  par_gap1,
  par_gap2
)
}
\arguments{
\item{get_cycle_edge_info}{input_net_file, output_path, res_path, input_tumor_name, input_tumor_data, input_normal_data
par_up1 = 1
par_up2 = 10
par_gap1 = -2
par_gap2 = 1.71}
}
\description{
This function for cycle get_cycle_edge_info
}
\examples{
get_cycle_edge_info(input_net_file, output_path, res_path, input_tumor_name, input_tumor_data, input_normal_data, par_up1, par_up2, par_gap1, par_gap2)

(max gene foldchange>(par_up1))&(max gene meanval>par_up2) -> up
(max gene foldchange<(par_gap1))|(max gene meanval<par_gap2) -> gap






}
\keyword{get_cycle_edge_info}
