% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_get_topo_info.R
\name{get_struct_sort_info}
\alias{get_struct_sort_info}
\title{get_struct_sort_info function}
\usage{
get_struct_sort_info(
  input_net_file,
  output_path,
  res_path,
  hub_para1 = 5,
  hub_para2 = 5,
  hub_para3 = 3,
  web_para1 = 0.18,
  web_para2 = 3,
  web_para3 = 7,
  web_para4 = 2,
  web_para5 = 10,
  idv_para1 = 0.16,
  idv_para2 = 1,
  idv_para3 = 2,
  idv_para4 = 20,
  idv_para5 = 0.56,
  idv_para6 = 20
)
}
\arguments{
\item{get_struct_sort_info}{input_net_file output_path res_path
hub_para1 = 5,
hub_para2 = 5,
hub_para3 = 3,
web_para1 = .18,
web_para2 = 3,
web_para3 = 7,
web_para4 = 2,
web_para5 = 10,
idv_para1 = .16,
idv_para2 = 1,
idv_para3 = 2,
idv_para4 = 20,
idv_para5 = .56,
idv_para6 = 20,}
}
\description{
This function for cycle get_struct_sort_info
}
\examples{
get_struct_sort_info()


}
\keyword{get_struct_sort_info}
