% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getBasicCycle}
\alias{getBasicCycle}
\title{getBasicCycle function}
\usage{
getBasicCycle(input_net_file)
}
\description{
Build a map based on input_net_file, search for basic cycles, and calculate the expression information of these rings
}
\examples{
getBasicCycle(input_net_file)


}
\keyword{getBasicCycle}
