% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getCycleFlux}
\alias{getCycleFlux}
\title{getCycleFlux function}
\usage{
getCycleFlux(
  basic_cycle_result,
  input_deg_gene_file,
  single_graph = FALSE,
  net_graph = FALSE
)
}
\arguments{
\item{single_graph:}{If this parameter is TRUE, an image will be generated in this directory for each cycle that matches the shift definition.}

\item{net_graph:}{If this parameter is TRUE, an image containing all cycles that match the shift definition will be generated in this directory.
Default:
single_graph=FALSE
net_graph=FALSE}
}
\description{
This function for cycle shift situation
}
\details{
For an edge, if any gene on the edge is up, the edge is up, if the edge contains both the gap gene and the up gene, it is regarded as a gap.
}
\examples{
getCycleFlux(basic_cycle_result, input_deg_gene_file, FALSE, FALSE)

}
\keyword{getCycleFlux}
